Update and rename Main.java to ListaPropositos.java
Se agregó la sentencia de repetición for, para condicionar la cantidad de propósitos que puede guardar el usuario.
Se cambio la sentencia condicional, de control de flujo, switch por if, por cuanto es la más adecuada al siguiente ejemplo.
